flask-example/flask-mysql/flask-mysql-simple-1-fail/requirements.txt
Flask
flask-mysql-connector
python-dotenv